feat: show onchain options in receive screen
